+++ src/pages/index.js
@@ -7,4 +7,3 @@ export { default as LoginPage } from './LoginPage'
 export { default as Signup } from './Signup'
 export { default as SingleApi } from './SingleApi'
 export { default as UserProfile } from './UserProfile'
-export { default as SignupPage } from './SignupPage'
